Refactor into shared server API module.
