+++ src/main/java/cloud/filibuster/junit/server/FilibusterServerAPI.java
@@ -1,0 +1,166 @@
+package cloud.filibuster.junit.server;
+
+import cloud.filibuster.instrumentation.datatypes.FilibusterExecutor;
+import cloud.filibuster.instrumentation.helpers.Response;
+import cloud.filibuster.junit.exceptions.FilibusterServerBadResponseException;
+import com.linecorp.armeria.client.WebClient;
+import com.linecorp.armeria.common.AggregatedHttpResponse;
+import com.linecorp.armeria.common.HttpHeaderNames;
+import com.linecorp.armeria.common.HttpMethod;
+import com.linecorp.armeria.common.RequestHeaders;
+import com.linecorp.armeria.common.ResponseHeaders;
+import org.json.JSONObject;
+
+import java.util.Objects;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+
+public class FilibusterServerAPI {
+    public static void analysisFile(WebClient webClient, JSONObject jsonAnalysisConfiguration) throws ExecutionException, InterruptedException {
+        CompletableFuture<Void> analysisFileFuture = CompletableFuture.supplyAsync(() -> {
+            RequestHeaders postJson = RequestHeaders.of(
+                    HttpMethod.POST,
+                    "/filibuster/analysis-file",
+                    HttpHeaderNames.CONTENT_TYPE,
+                    "application/json",
+                    "X-Filibuster-Instrumentation",
+                    "true");
+            AggregatedHttpResponse response = webClient.execute(postJson, jsonAnalysisConfiguration.toString()).aggregate().join();
+            ResponseHeaders headers = response.headers();
+            String statusCode = headers.get(HttpHeaderNames.STATUS);
+
+            if (statusCode == null) {
+                FilibusterServerBadResponseException.logAndThrow("analysisFile, statusCode: null");
+            }
+
+            if (!Objects.equals(statusCode, "200")) {
+                FilibusterServerBadResponseException.logAndThrow("analysisFile, statusCode: " + statusCode);
+            }
+
+            return null;
+        }, FilibusterExecutor.getExecutorService());
+
+        analysisFileFuture.get();
+    }
+
+    public static void terminate(WebClient webClient) throws ExecutionException, InterruptedException {
+        CompletableFuture<Void> terminateFuture = CompletableFuture.supplyAsync(() -> {
+            RequestHeaders getJson = RequestHeaders.of(
+                    HttpMethod.GET,
+                    "/filibuster/terminate",
+                    HttpHeaderNames.CONTENT_TYPE,
+                    "application/json",
+                    "X-Filibuster-Instrumentation",
+                    "true");
+            AggregatedHttpResponse response = webClient.execute(getJson).aggregate().join();
+            ResponseHeaders headers = response.headers();
+            String statusCode = headers.get(HttpHeaderNames.STATUS);
+
+            if (statusCode == null) {
+                FilibusterServerBadResponseException.logAndThrow("terminate, statusCode: null");
+            }
+
+            if (!Objects.equals(statusCode, "200")) {
+                FilibusterServerBadResponseException.logAndThrow("terminate, statusCode: " + statusCode);
+            }
+
+            return null;
+        }, FilibusterExecutor.getExecutorService());
+
+        terminateFuture.get();
+    }
+
+    public static void teardownsCompleted(WebClient webClient, int currentIteration) throws ExecutionException, InterruptedException {
+        CompletableFuture<Void> teardownsCompletedFuture = CompletableFuture.supplyAsync(() -> {
+            RequestHeaders getJson = RequestHeaders.of(
+                    HttpMethod.GET,
+                    "/filibuster/teardowns-completed/" + currentIteration,
+                    HttpHeaderNames.CONTENT_TYPE,
+                    "application/json",
+                    "X-Filibuster-Instrumentation",
+                    "true");
+            AggregatedHttpResponse response = webClient.execute(getJson).aggregate().join();
+            ResponseHeaders headers = response.headers();
+            String statusCode = headers.get(HttpHeaderNames.STATUS);
+
+            if (statusCode == null) {
+                FilibusterServerBadResponseException.logAndThrow("teardownsCompleted, statusCode: null");
+            }
+
+            if (!Objects.equals(statusCode, "200")) {
+                FilibusterServerBadResponseException.logAndThrow("teardownsCompleted, statusCode: " + statusCode);
+            }
+
+            return null;
+        }, FilibusterExecutor.getExecutorService());
+
+        teardownsCompletedFuture.get();
+    }
+
+    public static void recordIterationComplete(WebClient webClient, int currentIteration, boolean exceptionOccurred) throws ExecutionException, InterruptedException {
+        int exceptionOccurredInt;
+
+        if (exceptionOccurred) {
+            exceptionOccurredInt = 1;
+        } else {
+            exceptionOccurredInt = 0;
+        }
+
+        CompletableFuture<Void> updateFuture = CompletableFuture.supplyAsync(() -> {
+            RequestHeaders postJson = RequestHeaders.of(
+                    HttpMethod.POST,
+                    "/filibuster/complete-iteration/" + currentIteration + "/exception/" + exceptionOccurredInt,
+                    HttpHeaderNames.CONTENT_TYPE,
+                    "application/json",
+                    "X-Filibuster-Instrumentation",
+                    "true");
+            AggregatedHttpResponse response = webClient.execute(postJson).aggregate().join();
+            ResponseHeaders headers = response.headers();
+            String statusCode = headers.get(HttpHeaderNames.STATUS);
+
+            if (statusCode == null) {
+                FilibusterServerBadResponseException.logAndThrow("recordIterationComplete, statusCode: null");
+            }
+
+            if (!Objects.equals(statusCode, "200")) {
+                FilibusterServerBadResponseException.logAndThrow("recordIterationComplete, statusCode: " + statusCode);
+            }
+
+            return null;
+        }, FilibusterExecutor.getExecutorService());
+
+        updateFuture.get();
+    }
+
+    public static boolean hasNextIteration(WebClient webClient, int currentIteration, String caller) throws ExecutionException, InterruptedException {
+        CompletableFuture<Boolean> hasNextIteration = CompletableFuture.supplyAsync(() -> {
+            try {
+                RequestHeaders getJson = RequestHeaders.of(
+                        HttpMethod.GET,
+                        "/filibuster/has-next-iteration/" + currentIteration + "/" + caller,
+                        HttpHeaderNames.ACCEPT,
+                        "application/json",
+                        "X-Filibuster-Instrumentation",
+                        "true");
+                AggregatedHttpResponse response = webClient.execute(getJson).aggregate().join();
+                ResponseHeaders headers = response.headers();
+                String statusCode = headers.get(HttpHeaderNames.STATUS);
+
+                if (statusCode == null) {
+                    FilibusterServerBadResponseException.logAndThrow("hasNextIteration, statusCode: null");
+                }
+
+                if (!Objects.equals(statusCode, "200")) {
+                    FilibusterServerBadResponseException.logAndThrow("hasNextIteration, statusCode: " + statusCode);
+                }
+
+                JSONObject jsonObject = Response.aggregatedHttpResponseToJsonObject(response);
+                return jsonObject.getBoolean("has-next-iteration");
+            } catch (RuntimeException e) { // TODO: do we need this?
+                return false;
+            }
+        }, FilibusterExecutor.getExecutorService());
+
+        return hasNextIteration.get();
+    }
+}
